+++ src/main/java/com/project/trainingdiary/service/TrainerService.java
@@ -13,9 +13,9 @@ import com.project.trainingdiary.exception.impl.PtContractNotExistException;
 import com.project.trainingdiary.exception.impl.TraineeNotExistException;
 import com.project.trainingdiary.exception.impl.TrainerNotFoundException;
 import com.project.trainingdiary.repository.InBodyRecordHistoryRepository;
-import com.project.trainingdiary.repository.PtContractRepository;
 import com.project.trainingdiary.repository.TraineeRepository;
 import com.project.trainingdiary.repository.TrainerRepository;
+import com.project.trainingdiary.repository.ptContract.PtContractRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
